[REFACTORING] Split IMAPServerAppendLimitTest
