+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -256,76 +256,6 @@ class IMAPServerTest {
 
 
 
-    @Nested
-    class Limit {
-        IMAPServer imapServer;
-        private int port;
-
-        @BeforeEach
-        void beforeEach() throws Exception {
-            imapServer = createImapServer("imapServer.xml");
-            port = imapServer.getListenAddresses().get(0).getPort();
-        }
-
-        @AfterEach
-        void tearDown() {
-            imapServer.destroy();
-        }
-
-        @Test
-        void smallAppendsShouldWork() throws Exception {
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", SMALL_MESSAGE))
-                .doesNotThrowAnyException();
-
-            assertThat(testIMAPClient.select("INBOX")
-                .readFirstMessage())
-                .contains("\r\n" + SMALL_MESSAGE + ")\r\n");
-        }
-
-        @Test
-        void mediumAppendsShouldWork() throws Exception {
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", _65K_MESSAGE))
-                .doesNotThrowAnyException();
-
-            assertThat(testIMAPClient.select("INBOX")
-                .readFirstMessage())
-                .contains("\r\n" + _65K_MESSAGE + ")\r\n");
-        }
-
-        @Test
-        void largeAppendsShouldBeRejected() {
-            assertThatThrownBy(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", _129K_MESSAGE))
-                .isInstanceOf(RuntimeException.class)
-                .hasMessageContaining("BAD APPEND failed.");
-        }
-
-        @Test
-        void capabilityAdvertizeAppendLimit() throws Exception {
-            assertThat(
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .capability())
-                .contains("APPENDLIMIT=131072");
-        }
-
-        @Test
-        void statusAdvertizeAppendLimit() throws Exception {
-            assertThat(
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .sendCommand("STATUS \"INBOX\" (APPENDLIMIT)"))
-                .contains("* STATUS \"INBOX\" (APPENDLIMIT 131072)");
-        }
-    }
 
     @Nested
     class AdminUsers {
